refactor: Update the angular cli project so that it can be addressed with the ng xx commands.
The previous commit left the ng xx commands broken because of some path issues in the
.angular-cli.json. In order to make this work there is a work-around having a symbolic link out to
the library source, which is required because of a limitation in Angular CLI and AOT static
analysis.
