--- e2e/app.po.ts
@@ -1,11 +1,0 @@
-import { browser, element, by } from 'protractor';
-
-export class Angular2MaterializePage {
-  navigateTo() {
-    return browser.get('/');
-  }
-
-  getParagraphText() {
-    return element(by.css('app-root h1')).getText();
-  }
-}
