--- e2e/tsconfig.e2e.json
@@ -1,12 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../out-tsc/e2e",
-    "module": "commonjs",
-    "target": "es5",
-    "types":[
-      "jasmine",
-      "node"
-    ]
-  }
-}
