+++ sample/karma.conf.js
