+++ sample/package.json
@@ -22,20 +22,25 @@
   },
   "homepage": "https://github.com/InfomediaLtd/angular2-materialize#readme",
   "scripts": {
+    "ng": "ng",
     "commit": "git-cz",
-    "build": "gulp"
-  },
-  "peerDependencies": {
-    "materialize-css": "^0.97.8"
+    "start": "ng serve",
+    "build": "ng build",
+    "postbuild": "cat dist/sample/index.html | sed 's/\"\\.\\./\"\\./g' > dist/index.html",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e",
+    "src-build": "gulp"
   },
   "config": {
     "commitizen": {
       "path": "node_modules/cz-conventional-changelog"
     },
     "ghooks": {
-      "pre-commit": "npm run build"
+      "pre-commit": "npm run src-build"
     }
   },
+  "private": true,
   "dependencies": {
     "@angular/common": "^4.0.0",
     "@angular/compiler": "^4.0.0",
@@ -46,25 +51,26 @@
     "@angular/platform-browser-dynamic": "^4.0.0",
     "@angular/router": "^4.0.0",
     "core-js": "^2.4.1",
-    "hammerjs": "^2.0.8",
-    "jquery": "^2.2.4",
-    "materialize-css": "^0.97.8",
     "rxjs": "^5.1.0",
     "zone.js": "^0.8.4"
   },
   "devDependencies": {
+    "@angular/cli": "1.0.0",
     "@angular/compiler-cli": "^4.0.0",
-    "commitizen": "2.9.6",
-    "cz-conventional-changelog": "2.0.0",
-    "ghooks": "2.0.0",
-    "gulp": "3.9.1",
-    "gulp-rimraf": "0.2.1",
-    "gulp-typescript": "3.1.6",
-    "materialize-css": "^0.97.8",
-    "run-sequence": "1.2.2",
-    "semantic-release": "6.3.2",
-    "typescript": "~2.2.2",
-    "webpack": "^2.4.1",
-    "webpack-dev-server": "^2.4.2"
+    "@types/jasmine": "2.5.38",
+    "@types/node": "~6.0.60",
+    "codelyzer": "~2.0.0",
+    "jasmine-core": "~2.5.2",
+    "jasmine-spec-reporter": "~3.2.0",
+    "karma": "~1.4.1",
+    "karma-chrome-launcher": "~2.0.0",
+    "karma-cli": "~1.0.1",
+    "karma-jasmine": "~1.1.0",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-coverage-istanbul-reporter": "^0.2.0",
+    "protractor": "~5.1.0",
+    "ts-node": "~2.0.0",
+    "tslint": "~4.5.0",
+    "typescript": "~2.2.0"
   }
 }
