+++ sample/src/app/components/dropdown.ts
@@ -1,4 +1,4 @@
-import {MaterializeDirective} from "../../../src/index";
+import {MaterializeDirective} from "../../../lib/index";
 import {Component} from "@angular/core"
 
 @Component({
