+++ sample/src/app/components/model-bindings/index.ts
