+++ sample/src/app/components/side-nav.ts
@@ -1,4 +1,4 @@
-import {MaterializeDirective} from "../../../src/index";
+import {MaterializeDirective} from "../../../lib/index";
 import {Component} from "@angular/core"
 import {Location} from '@angular/common';
