+++ sample/src/app/components/tabs-routing.ts
