+++ sample/src/favicon.ico
