+++ sample/src/typings.d.ts
